tile: migrate drawing code to picosystem crate
